Stable_diffusion/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.20.0
safetensors>=0.3.0
Pillow>=9.0.0
matplotlib>=3.5.0
numpy>=1.21.0
huggingface-hub>=0.15.0
